LPS-71164 Move logic to XMLCustomSQLFileCheck
